[ANDROID-135] 홈 화면에서 지금까지 목표 조회, 내가 접속했던 최신의 날짜 저장 (#118)
* [ANDROID-135] feat : 오늘날짜를 저장하도록하는 dataStore 코드 작성

* [ANDROID-135] feat : 날짜 저장, 업데이트하는 레포지토리 코드 작성

* [ANDROID-135] feat : 날짜 불러오고 저장하는 UseCase 작성

* [ANDROID-135] fix : 목표 완료 퍼센테이지를 double 타입으로 전달해주고있었다

* [ANDROID-135] feat : 홈 화면에서 목표 설정 정보를 받고 오늘날 정보를 로컬 DB에 저장하는 코드 작성

목표 완료 화면으로 이동하는것은 이 다음 브랜치에서 작업할 예정

+++ core/model/src/main/java/see/day/model/goal/CurrentGoal.kt
@@ -9,7 +9,7 @@ data class CurrentGoal(
     val startDate: String,
     val endDate: String,
     val completedDays: Int,
-    val achievementRate: Int,
+    val achievementRate: Double,
     val treeStage: TreeStage,
     val canCreateNew: Boolean
 )
